requirements.txt
streamlit==1.37.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==2.0.1
plotly==5.23.0
python-dotenv==1.0.1